beef up slimmns: fix -W, it was not ready to deal with the nature of slash file systems W.R.T. hardlinked directory traversal; add a -i flag for initializing filesystems for use with I/O nodes
git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@8645 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ slimmns/wipedir.c
@@ -5,6 +5,9 @@
 #include <sys/types.h>
 #include <sys/stat.h>
 
+#include <dirent.h>
+#include <err.h>
+#include <errno.h>
 #include <fts.h>
 #include <stdio.h>
 #include <string.h>
@@ -12,26 +15,89 @@
 
 #include "psc_util/log.h"
 
+#include "pathnames.h"
+#include "mkfn.h"
+
+extern int ion;
+
 void
-wipedir(const char *dir)
+wipefs(const char *dir)
 {
-	char * const pathv[] = { (char *)dir, NULL };
+	char *pathv[] = { (char *)dir, NULL };
+	char fn[PATH_MAX];
 	FTSENT *f;
 	FTS *fp;
 
+	/* first remove the user namespace */
 	/* XXX security implications of FTS_NOCHDIR? */
 	fp = fts_open(pathv, FTS_COMFOLLOW | FTS_NOCHDIR | FTS_PHYSICAL,
 	    NULL);
 	if (fp == NULL)
 		psc_fatal("fts_open %s", dir);
 	while ((f = fts_read(fp)) != NULL) {
+		if (f->fts_level < 1)
+			continue;
 		if (S_ISDIR(f->fts_statp->st_mode)) {
-			if ((f->fts_info & FTS_DP) &&
-			    strcmp(f->fts_path, dir) &&
+			/* skip slash namespace directories */
+			if (strncmp(f->fts_name, SL_PATH_PREFIX,
+			    strlen(SL_PATH_PREFIX)) == 0)
+				fts_set(fp, f, FTS_SKIP);
+			else if ((f->fts_info & FTS_DP) &&
 			    rmdir(f->fts_path) == -1)
 				psc_fatal("rmdir %s", f->fts_path);
 		} else if (unlink(f->fts_path) == -1)
-			psc_fatal("unlink %s", f->fts_path);
+			psc_error("unlink %s", f->fts_path);
+	}
+	fts_close(fp);
+
+	/* remove the slash fid namespace */
+	xmkfn(fn, "%s/%s", dir, SL_PATH_FIDNS);
+	pathv[0] = fn;
+
+	/* XXX security implications of FTS_NOCHDIR? */
+	fp = fts_open(pathv, FTS_COMFOLLOW | FTS_NOCHDIR | FTS_PHYSICAL,
+	    NULL);
+	if (fp == NULL)
+		psc_fatal("fts_open %s", fn);
+	while ((f = fts_read(fp)) != NULL) {
+		if (f->fts_level < 1)
+			continue;
+		if (S_ISDIR(f->fts_statp->st_mode)) {
+			if ((f->fts_info & FTS_DP) &&
+			    rmdir(f->fts_path) == -1)
+				psc_fatal("rmdir %s", f->fts_path);
+			/* do not descend into hardlinked directories */
+			else if (f->fts_level > 3)
+				fts_set(fp, f, FTS_SKIP);
+		} else if (unlink(f->fts_path) == -1)
+			psc_error("unlink %s", f->fts_path);
+	}
+	fts_close(fp);
+
+	if (ion)
+		return;
+
+	/* remove the slash replication queue */
+	xmkfn(fn, "%s/%s", dir, SL_PATH_REPLS);
+	pathv[0] = fn;
+
+	/* XXX security implications of FTS_NOCHDIR? */
+	fp = fts_open(pathv, FTS_COMFOLLOW | FTS_NOCHDIR | FTS_PHYSICAL,
+	    NULL);
+	if (fp == NULL)
+		psc_fatal("fts_open %s", fn);
+	while ((f = fts_read(fp)) != NULL) {
+		if (f->fts_level < 1)
+			continue;
+		if (S_ISDIR(f->fts_statp->st_mode)) {
+			if ((f->fts_info & FTS_DP) &&
+			    rmdir(f->fts_path) == -1)
+				psc_fatal("rmdir %s", f->fts_path);
+			/* do not descend into hardlinked directories */
+			else if (f->fts_level > 1)
+				fts_set(fp, f, FTS_SKIP);
+		} else if (unlink(f->fts_path) == -1)
+			psc_error("unlink %s", f->fts_path);
 	}
 	fts_close(fp);
 }
